trying something
In the other version of this code, I define a new bragg_theta for the peak I analyze, then use that to define the q1 vector. Here, instead I use the offset in q1 from the calibrated q1 in the definition for q1. In this whey I am skipping the approximation of choosing the bragg peak center which is even hard to define what that is.